refactor: refactor after merge
