+++ src/app/auth/guards/access.guard.ts
@@ -1,9 +1,8 @@
 import { inject } from '@angular/core';
-import { Router } from '@angular/router';
+import { ActivatedRouteSnapshot, CanActivateFn, createUrlTreeFromSnapshot, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
-import { UserActions } from '@app/core/store/user-store/actions/user.actions';
-import { selectUserRole } from '@app/core/store/user-store/selectors/user.selectors';
-import { of, switchMap } from 'rxjs';
+import { selectUser, selectUserLoading, selectUserRole } from '@app/core/store/user-store/selectors/user.selectors';
+import { filter, map, of, switchMap } from 'rxjs';
 import { AuthService } from '../services/auth.service';
 
 export const canActiveAuth = () => {
@@ -23,8 +22,6 @@ export const canActiveAdmin = () => {
   const router = inject(Router);
   const store = inject(Store);
 
-  store.dispatch(UserActions.loadUser());
-
   return store.select(selectUserRole).pipe(
     switchMap((role) => {
       if (role === 'manager') {
@@ -33,5 +30,16 @@ export const canActiveAdmin = () => {
 
       return of(router.createUrlTree(['not-found']));
     })
+  );
+};
+
+export const authGuard: CanActivateFn = (route: ActivatedRouteSnapshot) => {
+  const store = inject(Store);
+
+  return store.select(selectUserLoading).pipe(
+    filter((isUserLoading) => !isUserLoading),
+    switchMap(() =>
+      store.select(selectUser).pipe(map((user) => !!user || createUrlTreeFromSnapshot(route, ['/signin'])))
+    )
   );
 };
